NVIDIA: SAUCE: phy: tegra: Increase timeout for USB2_TRK_COMPLETED polling
BugLink: https://bugs.launchpad.net/bugs/2080908

The timeout for polling on USB2_TRK_COMPLETED in the UTMI bias pad
power-on sequence has been increased from 100 to 200 cycles.

This change is necessary due to the increased number of cycles required
to settle all biases in the new circuits. Additionally, the clock speed
of the tracking circuit has been reduced to 1 MHz, further contributing
to the need for a longer timeout.

http://nvbugs/4796326

Signed-off-by: Wayne Chang <waynec@nvidia.com>
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>